+++ void_logging/api/rewards/__init__.py
@@ -2,8 +2,7 @@
 
 from .log import Log
 from .logged_reward import LoggedReward
-from .logged_combined_reward import LoggedCombinedReward
 from .reward_logger import RewardLogger
 from .logged_float import Logged
 
-__all__ = ["Log", "LoggedReward", "Logged", "LoggedCombinedReward", "RewardLogger"]
+__all__ = ["Log", "LoggedReward", "Logged", "RewardLogger"]