+++ void_logging/rlgym_learn/multi_logger.py
@@ -1,41 +1,35 @@
 """Module for the multi logger"""
 
 from dataclasses import dataclass
-from typing import Generic, Dict, Any, List
+from typing import Any, Dict, Generic, List
 
-from pydantic import BaseModel
-from rlgym_learn.api.agent_controller import AgentControllerData
+from pydantic import BaseModel, ValidationError
+from rlgym_learn.api.typing import AgentControllerData
 from rlgym_learn_algos.logging import (
-    InnerMetricsLoggerConfig,
-    InnerMetricsLoggerAdditionalDerivedConfig,
     DictMetricsLogger,
+    InnerMetricsLoggerAdditionalDerivedConfig,
+    InnerMetricsLoggerConfig,
 )
+from rlgym_learn_algos.logging.metrics_logger import DerivedMetricsLoggerConfig
 
-from ..logging_utils import print_metrics
 
-
-class MultiLoggerConfigModel(BaseModel):
-    "A config class for the multi logger in case i want to add config"
+class MultiLoggerConfigModel(BaseModel, Generic[InnerMetricsLoggerConfig]):
+    inner_metrics_logger_config: List[InnerMetricsLoggerConfig | None] = []
 
 
 @dataclass
 class MultiLoggerAdditionalDerivedConfig(
-    Generic[InnerMetricsLoggerConfig, InnerMetricsLoggerAdditionalDerivedConfig]
+    Generic[InnerMetricsLoggerAdditionalDerivedConfig]
 ):
-    """Nah i'm not sure what that class does, but it's what the WandB logger was doing"""
-
-    inner_metrics_logger_config: InnerMetricsLoggerConfig | None = None
-    inner_metrics_logger_additional_derived_config: (
+    inner_metrics_logger_additional_derived_config: List[
         InnerMetricsLoggerAdditionalDerivedConfig | None
-    ) = None
+    ] = []
 
 
 class MultiLogger(
     DictMetricsLogger[
-        MultiLoggerConfigModel,
-        MultiLoggerAdditionalDerivedConfig[
-            InnerMetricsLoggerConfig, InnerMetricsLoggerAdditionalDerivedConfig
-        ],
+        MultiLoggerConfigModel[InnerMetricsLoggerConfig],
+        MultiLoggerAdditionalDerivedConfig[InnerMetricsLoggerAdditionalDerivedConfig],
         AgentControllerData,
     ],
     Generic[
@@ -52,32 +46,80 @@ class MultiLogger(
         self.agent_metrics = {}
 
     def validate_config(self, config_obj: Dict[str, Any]) -> MultiLoggerConfigModel:
-        return MultiLoggerConfigModel.model_validate(config_obj)
+        _base_config = MultiLoggerConfigModel.model_validate(config_obj)
+
+        # This should never occur
+        if "inner_metrics_logger_config" not in config_obj:
+            raise ValidationError(
+                "Expected 'inner_metrics_logger_config' in config object"
+            )
+
+        try:
+            _len = len(config_obj["inner_metrics_logger_config"])
+
+            if _len != len(self.inner_metrics_loggers):
+                raise ValidationError(
+                    f"You gave {_len} config objects for {len(self.inner_metrics_loggers)}. Please provide an equal number for both. If one of the metric loggers used doesn't need any config, you can use None as a filler"
+                )
+        except TypeError as e:
+            raise TypeError(
+                f"The config you gave to the multi logger has no length. Config type: {type(config_obj['inner_metrics_logger_config'])}"
+            ) from e
+
+        for _idx, inner_metric_logger in enumerate(self.inner_metrics_loggers):
+            _config = inner_metric_logger.validate_config(
+                config_obj["inner_metrics_logger_config"][_idx]
+            )
+
+            _base_config.inner_metrics_logger_config[_idx] = _config
+
+        return _base_config
+
+    def load(
+        self,
+        config: DerivedMetricsLoggerConfig[
+            MultiLoggerConfigModel[InnerMetricsLoggerConfig],
+            MultiLoggerAdditionalDerivedConfig[
+                InnerMetricsLoggerAdditionalDerivedConfig
+            ],
+        ],
+    ):
+        self.config = config
+
+        for _idx, inner_metric_logger in enumerate(self.inner_metrics_loggers):
+            _inner_derived_config = config.additional_derived_config.inner_metrics_logger_additional_derived_config[
+                _idx
+            ]
+            _inner_config = config.metrics_logger_config.inner_metrics_logger_config[
+                _idx
+            ]
+
+            if _inner_config is not None and _inner_derived_config is not None:
+                inner_metric_logger.load(
+                    DerivedMetricsLoggerConfig(
+                        metrics_logger_config=_inner_config,
+                        additional_derived_config=_inner_derived_config,
+                        checkpoint_load_folder=config.checkpoint_load_folder,
+                        agent_controller_name=config.agent_controller_name,
+                    )
+                )
 
     def collect_env_metrics(self, data: List[Dict[str, Any]]):
-        env_metrics = {}
-
         for inner_metric_logger in self.inner_metrics_loggers:
             inner_metric_logger.collect_env_metrics(data)
 
-            if hasattr(inner_metric_logger, "env_metrics"):
-                env_metrics = env_metrics | inner_metric_logger.env_metrics
-
-        self.env_metrics = env_metrics
-
     def collect_agent_metrics(self, data: AgentControllerData):
-        agent_metrics = {}
-
         for inner_metric_logger in self.inner_metrics_loggers:
             inner_metric_logger.collect_agent_metrics(data)
 
-            if hasattr(inner_metric_logger, "agent_metrics"):
-                agent_metrics = agent_metrics | inner_metric_logger.agent_metrics
-
-        self.agent_metrics = agent_metrics
-
     def get_metrics(self) -> Dict[str, Any]:
-        return self.env_metrics | self.agent_metrics
+        _metrics = {}
+
+        for inner_metric_logger in self.inner_metrics_loggers:
+            _metrics = _metrics | inner_metric_logger.get_metrics()
+
+        return _metrics
 
     def report_metrics(self):
-        print_metrics(self.get_metrics())
+        for inner_metric_logger in self.inner_metrics_loggers:
+            inner_metric_logger.report_metrics()